Фикс http://forum.ss13.ru/index.php?showtopic=11331&st=840#entry396966

+++ code/modules/janitor/holosign.dm
@@ -17,7 +17,8 @@
 	if(flag)
 		var/turf/T = get_turf(target)
 		var/obj/effect/overlay/holograph/H = locate() in T
-		if (istype(target, /obj/structure/janitorialcart))	return
+		if(istype(target, /obj/structure/janitorialcart))	return
+		if(istype(target, /obj/item/weapon/storage)) 		return
 		else if(H)
 			user << "<span class='notice'>You use [src] to destroy [H].</span>"
 			qdel(H)
